I think I fixed it
***I still don’t know why there was that automatic blank space at the
top which we noticed after class.  After inspecting the element a
million times I found it was something between the HTML tag and head
tag…didn’t figure out what was going on but a 1px padding at the top
meant that the space would be styled like the wrapper which makes it
look normal.  If you pay close attention it looks a bit off but I can’t
figure out what else to do.